Separate the lang service and arduino IDE funtionalities.
